Add the ability to download and unpack all the extensions

+++ script/extension.ts
@@ -3,6 +3,7 @@ import { readFile, rmdir, writeFile } from 'node:fs/promises';
 import { downloadAndSave, listExtensions, uploadPattern } from './lib/extension';
 import path from 'node:path';
 import { execSync } from 'node:child_process';
+import { existsSync } from 'node:fs';
 
 const extensionNameMap = new Map([
 ]);
@@ -29,22 +30,47 @@ program.command('list')
 program.command('download')
     .description('Downloads the AUX for the given extension to the dist folder.')
     .argument('<name>', 'The name of the extension to download.')
-    .action(async (name, options) => {
-        await downloadAndSave(name, options.version);
+    .action(async (name) => {
+        await downloadAndSave(name);
     });
+
+program.command('download-all')
+    .description('Downloads all AUXes for the given extension to the dist folder.')
+    .action(async () => {
+        const list = await listExtensions();
+        for (const ext of list) {
+            await downloadAndSave(ext.name);
+        }
+    });
+
+
+async function unpack(name: string) {
+    const { filePath, ...ext } = await downloadAndSave(name, extensionNameMap.get(name) || `${name}.aux`);
+    const extensionPath = path.resolve('packages', extensionNameMap.get(name) || name);
+    if (existsSync(extensionPath)) {
+        await rmdir(extensionPath, { recursive: true });
+    }
+    execSync(`casualos unpack-aux --overwrite "${filePath}" ./packages`, { stdio: 'ignore' });
+
+    const extensionJsonPath = path.resolve(extensionPath, 'extension.json');
+    await writeFile(extensionJsonPath, JSON.stringify(ext.meta, null, 2), 'utf-8');
+    console.log(`Unpacked extension ${name} to packages folder.`);
+}
 
 program.command('unpack')
     .description('Downloads and unpacks the AUX for the given extension into the packages folder.')
     .argument('<name>', 'The name of the extension to download.')
-    .action(async (name, options) => {
-        const { filePath, ...ext } = await downloadAndSave(name, options.version, extensionNameMap.get(name) || `${name}.aux`);
-        const extensionPath = path.resolve('packages', extensionNameMap.get(name) || name);
-        await rmdir(extensionPath, { recursive: true });
-        execSync(`casualos unpack-aux --overwrite "${filePath}" ./packages`, { stdio: 'ignore' });
+    .action(async (name) => {
+        await unpack(name);
+    });
 
-        const extensionJsonPath = path.resolve(extensionPath, 'extension.json');
-        await writeFile(extensionJsonPath, JSON.stringify(ext.meta, null, 2), 'utf-8');
-        console.log(`Unpacked extension ${name} to packages folder.`);
+program.command('unpack-all')
+    .description('Downloads and unpacks the AUX for all extensions into the packages folder.')
+    .action(async () => {
+        const list = await listExtensions();
+        for (const ext of list) {
+            await unpack(ext.name);
+        }
     });
 
 // program.command('upload')
